pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "euroeval-docker-runner"
version = "0.1.0"
description = "Dockerized EuroEval runner with OpenAI-compatible API support and pre-cached datasets"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "euroeval[all]>=15.0.0",  # keep relatively fresh
  "rich>=13.7.1",
  "typer>=0.12.3",
  "prompt-toolkit>=3.0.47",
  "python-dotenv>=1.0.1",
  "termcolor2>=0.0.1" 
]

[project.scripts]
euroeval-runner = "euroeval_runner.cli:app"
